backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
sqlalchemy>=2.0.9
alembic>=1.10.3
psycopg2-binary>=2.9.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.1.0
pillow>=9.5.0
celery>=5.2.7
redis>=4.5.4
httpx>=0.24.0
jinja2>=3.1.2
pytest>=7.3.1
pytest-asyncio>=0.21.0
